Refractor SongList
