+++ packages/client/src/components/playing/listen/History.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react";
 import { useResource } from "react-request-hook";
 import api from "../../../core/api/model";
 import useReload from "../../../core/reloadHook";
-import { DefaultSongEntryList } from "../snippets/SongList";
+import { DefaultSongEntryList } from "../snippets/songlist/SongList";
 import ScreenNavigation from "../../util/ScreenNavigation";
 import useResourceWithPermission from "../../../core/api/permissionWrapperHook";
 import { Permission, SongEntry } from "../../../core/types";
